feat: contract initialization and deployment
