+++ .env.example
@@ -1,18 +1,8 @@
-FUJI_PRIVATE_KEY=
 ALCHEMY_API_KEY=
 ETHERSCAN_API_KEY=
-SNOWTRACE_API_KEY=
-COINMARKETCAP_API_KEY=
-
-# https://datahub.figment.io/
-DATAHUB_API_KEY=
-
-# Needed if you are manually pushing a Docker image. But it's unlikely you will, as this is done automatically by GitHub Actions.
-DOCKER_USER=replace
-DOCKER_PASS=replace
+GOERLI_RPC_ENDPOINT=
 
 REPORT_GAS=true
 EVM_LOGGING=false
 
 DEPLOYER_PK=
-TOKEN_OWNER_PK=
