--- scripts/constants.ts
@@ -1,17 +1,0 @@
-export const CONTRACTS: Record<string, string> = {
-    escrow: "Escrow",
-    authority: "Authority",
-    treasury: "Treasury",
-    erc20Mock: "BanyanERC20Mock",
-    erc721Mock: "ERC721Mock",
-    erc1155Mock: "ERC1155Mock",
-};
-
-export const CONFIGURATION: Record<string, any> = {
-    hardhatChainId: 1337,
-};
-
-export const CONSTANTS: Record<string, any> = {
-    ZERO_ADDRESS: "0x0000000000000000000000000000000000000000",
-    ZERO_BYTES32: "0x0000000000000000000000000000000000000000000000000000000000000000",
-};
