--- scripts/deploy/02_deploy_treasury.ts
@@ -1,23 +1,0 @@
-import { HardhatRuntimeEnvironment } from "hardhat/types";
-import { DeployFunction } from "hardhat-deploy/types";
-import { CONTRACTS } from "../constants";
-import { Treasury } from '../../types/contracts/Treasury';
-
-const func: DeployFunction = async (hre: HardhatRuntimeEnvironment) => {
-    const { deployments, getNamedAccounts } = hre;
-    const { deploy } = deployments;
-    const { deployer } = await getNamedAccounts();
-
-    const authorityDeployment = await deployments.get(CONTRACTS.authority);
-
-    await deploy(CONTRACTS.treasury, {
-        from: deployer,
-        args: [authorityDeployment.address],
-        log: true,
-        skipIfAlreadyDeployed: true,
-    });
-};
-
-func.tags = [CONTRACTS.treasury, "migration", "production"];
-
-export default func;
