--- scripts/deploy/03_deploy_escrow.ts
@@ -1,24 +1,0 @@
-import { HardhatRuntimeEnvironment } from "hardhat/types";
-import { DeployFunction } from "hardhat-deploy/types";
-import { CONTRACTS, CONSTANTS } from "../constants";
-
-
-const func: DeployFunction = async (hre: HardhatRuntimeEnvironment) => {
-    const { deployments, getNamedAccounts } = hre;
-    const { deploy } = deployments;
-    const { deployer } = await getNamedAccounts();
-
-    const authorityDeployment = await deployments.get(CONTRACTS.authority);
-
-    await deploy(CONTRACTS.escrow, {
-        from: deployer,
-        args: [authorityDeployment.address, CONSTANTS.ZERO_ADDRESS],
-        log: true,
-        skipIfAlreadyDeployed: true,
-    });
-};
-
-func.tags = [CONTRACTS.escrow];
-func.dependencies = [CONTRACTS.authority, CONTRACTS.treasury];
-
-export default func;
