--- scripts/txHelper.ts
@@ -1,5 +1,0 @@
-import { ContractReceipt, ContractTransaction } from "@ethersproject/contracts";
-
-export async function waitFor(txPromise: Promise<ContractTransaction>): Promise<ContractReceipt> {
-    return await txPromise.then((tx) => tx.wait());
-}
